+++ yaml.go
@@ -27,11 +27,11 @@ import (
 	"io/ioutil"
 	"os"
 	"reflect"
-	"runtime"
 	"strconv"
 	"strings"
 
 	"github.com/pkg/errors"
+	"golang.org/x/text/transform"
 	"gopkg.in/yaml.v2"
 )
 
@@ -44,7 +44,7 @@ var (
 	_emptyDefault = `""`
 )
 
-func newYAMLProviderCore(files ...io.ReadCloser) (*yamlConfigProvider, error) {
+func newYAMLProviderCore(files ...io.Reader) (*yamlConfigProvider, error) {
 	var root interface{}
 	for _, v := range files {
 		var curr interface{}
@@ -134,33 +134,69 @@ func mergeMaps(dst interface{}, src interface{}) (interface{}, error) {
 // file names. All the objects are going to be merged and arrays/values
 // overridden in the order of the files.
 func NewYAMLProviderFromFiles(files ...string) (Provider, error) {
-	readers, err := filesToReaders(files...)
+	readClosers, err := filesToReaders(files...)
 	if err != nil {
 		return nil, err
 	}
 
-	p, err := newYAMLProviderCore(readers...)
-	if err != nil {
-		return nil, err
+	readers := make([]io.Reader, len(readClosers))
+	for i, r := range readClosers {
+		readers[i] = r
 	}
 
-	return newCachedProvider(p)
+	provider, err := newYAMLProviderFromReader(readers...)
+
+	for _, r := range readClosers {
+		nerr := r.Close()
+		if err == nil {
+			err = nerr
+		}
+	}
+
+	return provider, err
 }
 
 // NewYAMLProviderWithExpand creates a configuration provider from a set of YAML
 // file names with ${var} or $var values replaced based on the mapping function.
+// Variable names not wrapped in curly braces will be parsed according
+// to the shell variable naming rules:
+//
+//     ...a word consisting solely of underscores, digits, and
+//     alphabetics from the portable character set. The first
+//     character of a name is not a digit.
+//
+// For variables wrapped in braces, all characters between '{' and '}'
+// will be passed to the expand function.  e.g. "${foo:13}" will cause
+// "foo:13" to be passed to the expand function.  The sequence '$$' will
+// be replaced by a literal '$'.  All other sequences will be ignored
+// for expansion purposes.
 func NewYAMLProviderWithExpand(mapping func(string) (string, bool), files ...string) (Provider, error) {
-	readers, err := filesToReaders(files...)
+	readClosers, err := filesToReaders(files...)
 	if err != nil {
 		return nil, err
 	}
 
-	return newYAMLProviderFromReaderWithExpand(mapping, readers...)
+	readers := make([]io.Reader, len(readClosers))
+	for i, r := range readClosers {
+		readers[i] = r
+	}
+
+	provider, err := newYAMLProviderFromReaderWithExpand(mapping,
+		readers...)
+
+	for _, r := range readClosers {
+		nerr := r.Close()
+		if err == nil {
+			err = nerr
+		}
+	}
+
+	return provider, err
 }
 
-// NewYAMLProviderFromReader creates a configuration provider from a list of io.ReadClosers.
+// NewYAMLProviderFromReader creates a configuration provider from a list of io.Readers.
 // As above, all the objects are going to be merged and arrays/values overridden in the order of the files.
-func newYAMLProviderFromReader(readers ...io.ReadCloser) (Provider, error) {
+func newYAMLProviderFromReader(readers ...io.Reader) (Provider, error) {
 	p, err := newYAMLProviderCore(readers...)
 	if err != nil {
 		return nil, err
@@ -170,38 +206,34 @@ func newYAMLProviderFromReader(readers ...io.ReadCloser) (Provider, error) {
 }
 
 // NewYAMLProviderFromReaderWithExpand creates a configuration provider from
-// a list of `io.ReadClosers and uses the mapping function to expand values
+// a list of `io.Readers and uses the mapping function to expand values
 // in the underlying provider.
 func newYAMLProviderFromReaderWithExpand(
 	mapping func(string) (string, bool),
-	readers ...io.ReadCloser) (Provider, error) {
-	p, err := newYAMLProviderCore(readers...)
-	if err != nil {
-		return nil, err
+	readers ...io.Reader) (Provider, error) {
+
+	expandFunc := replace(mapping)
+
+	ereaders := make([]io.Reader, len(readers))
+	for i, reader := range readers {
+		ereaders[i] = transform.NewReader(
+			reader,
+			&expandTransformer{expand: expandFunc})
 	}
 
-	if err := p.root.applyOnAllNodes(replace(mapping)); err != nil {
-		return nil, err
-	}
-
-	return newCachedProvider(p)
+	return newYAMLProviderFromReader(ereaders...)
 }
 
 // NewYAMLProviderFromBytes creates a config provider from a byte-backed YAML
 // blobs. As above, all the objects are going to be merged and arrays/values
 // overridden in the order of the yamls.
 func NewYAMLProviderFromBytes(yamls ...[]byte) (Provider, error) {
-	closers := make([]io.ReadCloser, len(yamls))
+	readers := make([]io.Reader, len(yamls))
 	for i, yml := range yamls {
-		closers[i] = ioutil.NopCloser(bytes.NewReader(yml))
+		readers[i] = bytes.NewReader(yml)
 	}
 
-	p, err := newYAMLProviderCore(closers...)
-	if err != nil {
-		return nil, err
-	}
-
-	return newCachedProvider(p)
+	return newYAMLProviderFromReader(readers...)
 }
 
 func filesToReaders(files ...string) ([]io.ReadCloser, error) {
@@ -210,6 +242,9 @@ func filesToReaders(files ...string) ([]io.ReadCloser, error) {
 
 	for _, v := range files {
 		if reader, err := os.Open(v); err != nil {
+			for _, r := range readers {
+				r.Close()
+			}
 			return nil, err
 		} else if reader != nil {
 			readers = append(readers, reader)
@@ -326,64 +361,13 @@ func (n *yamlNode) Children() []*yamlNode {
 	return nodes
 }
 
-// Apply expand to all nested elements of a node.
-// There is no need to use reflection, because YAML unmarshaler is using
-// map[interface{}]interface{} to store objects and []interface{}
-// to store collections.
-func recursiveApply(node interface{}, expand func(string) string) interface{} {
-	if node == nil {
-		return nil
-	}
-	switch t := node.(type) {
-	case map[interface{}]interface{}:
-		for k := range t {
-			t[k] = recursiveApply(t[k], expand)
-		}
-		return t
-	case []interface{}:
-		for i := range t {
-			t[i] = recursiveApply(t[i], expand)
-		}
-		return t
-	}
-
-	return os.Expand(fmt.Sprint(node), expand)
-}
-
-func (n *yamlNode) applyOnAllNodes(expand func(string) string) (err error) {
-
-	defer func() {
-		if r := recover(); r != nil {
-			if _, ok := r.(runtime.Error); ok {
-				panic(r)
-			}
-
-			err = r.(error)
-		}
-	}()
-
-	n.value = recursiveApply(n.value, expand)
-
-	for _, c := range n.Children() {
-		if err := c.applyOnAllNodes(expand); err != nil {
-			return err
-		}
-	}
-
-	return
-}
-
-func unmarshalYAMLValue(reader io.ReadCloser, value interface{}) error {
+func unmarshalYAMLValue(reader io.Reader, value interface{}) error {
 	raw, err := ioutil.ReadAll(reader)
 	if err != nil {
 		return errors.Wrap(err, "failed to read the yaml config")
 	}
 
-	if err = yaml.Unmarshal(raw, value); err != nil {
-		return err
-	}
-
-	return reader.Close()
+	return yaml.Unmarshal(raw, value)
 }
 
 // Function to expand environment variables in returned values that have form: ${ENV_VAR:DEFAULT_VALUE}.
@@ -396,10 +380,8 @@ func unmarshalYAMLValue(reader io.ReadCloser, value interface{}) error {
 //
 // In the case that HTTP_PORT is not provided, default value (in this case 8080)
 // will be used.
-//
-// TODO: what if someone wanted a literal ${FOO} in config? need a small escape hatch
-func replace(lookUp func(string) (string, bool)) func(in string) string {
-	return func(in string) string {
+func replace(lookUp func(string) (string, bool)) func(in string) (string, error) {
+	return func(in string) (string, error) {
 		sep := strings.Index(in, _envSeparator)
 		var key string
 		var def string
@@ -414,16 +396,16 @@ func replace(lookUp func(string) (string, bool)) func(in string) string {
 		}
 
 		if envVal, ok := lookUp(key); ok {
-			return envVal
+			return envVal, nil
 		}
 
 		if def == "" {
-			panic(fmt.Errorf(`default is empty for %q (use "" for empty string)`, key))
+			return "", fmt.Errorf(`default is empty for %q (use "" for empty string)`, key)
 		} else if def == _emptyDefault {
-			return ""
+			return "", nil
 		}
 
-		return def
+		return def, nil
 	}
 }